Delete demo comments after one week
Comments on the demo site are for the purpose of trying out the site. We
expect a lot of "test and "hello" type of comments, and the page will
look strange with hundreds of old test comments on it.

This commit adds a subcommand to the `carnival` executable to delete
stale comments from the demo site.